+++ src/main/java/marche/traitement/participant/ProducteurLaitier.java
@@ -22,7 +22,7 @@ public class ProducteurLaitier extends Fermier implements Visitable {
      * @param uniteDeProductions unités de productions dont dispose un fermier à sa création
      * @param solde Gain dont dispose un fermier lors de sa création
      */
-    public ProducteurLaitier(ArrayList<ProduitFermier> inventaire, Collection<UniteDeProduction> uniteDeProductions, float solde){
+    public ProducteurLaitier(ArrayList<ProduitFermier> inventaire, Collection<UniteDeProduction> uniteDeProductions, double solde){
         super(inventaire,uniteDeProductions,solde);
     }
